[REF] project,hr_timesheet: reorganize project/src/js files
task-3300649
