+++ addons/hr_timesheet/static/src/views/timesheet_graph/timesheet_graph_view.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { projectGraphView } from "@project/js/project_graph_view";
+import { projectGraphView } from "@project/views/project_task_graph/project_graph_view";
 import { hrTimesheetGraphModel } from "./timesheet_graph_model";
 import { registry } from "@web/core/registry";
