+++ addons/project/__manifest__.py
@@ -69,12 +69,6 @@
             'project/static/src/utils/**/*',
             'project/static/src/components/**/*',
             'project/static/src/views/**/*',
-            'project/static/src/js/project_graph_view.js',
-            'project/static/src/js/project_pivot_model.js',
-            'project/static/src/js/project_pivot_view.js',
-            'project/static/src/js/project_rating_graph_view.js',
-            'project/static/src/js/project_rating_pivot_view.js',
-            'project/static/src/js/project_task_kanban_examples.js',
             'project/static/src/js/tours/project.js',
             'project/static/src/scss/project_dashboard.scss',
             'project/static/src/scss/project_form.scss',
