+++ addons/project/static/src/components/project_task_priority_switch_field/project_task_priority_switch_field.js
