+++ addons/project/static/src/views/project_rating_graph/project_rating_graph_view.js
