+++ addons/project/static/src/views/project_rating_pivot/project_rating_pivot_view.js
