+++ addons/project/static/src/views/project_task_graph/project_graph_view.js
