+++ addons/project/static/src/views/project_task_pivot/project_pivot_view.js
